+++ InvoicesToGo/Models/Item.swift
@@ -9,12 +9,12 @@ import Foundation
 
 struct Item {
     var name: String
-    var quantity: Int
+    var quantity: Double
     var rate: Double
 
     init(dictionary: [String: Any]) {
         name = dictionary["name"] as? String ?? ""
-        quantity = dictionary["quantity"] as? Int ?? 0
+        quantity = dictionary["quantity"] as? Double ?? 0.0
         rate = dictionary["rate"] as? Double ?? 0.0
     }
 }
